+++ components/transaction/TransactionDetails.tsx
@@ -13,8 +13,8 @@ import Chips from "../Chips";
 import IconButton from "@material-ui/core/IconButton";
 
 import { useRouter } from "next/router";
-import gql from "graphql-tag";
-import { useQuery } from "@apollo/react-hooks";
+import gql from "@apollo/client";
+import { useQuery } from "@apollo/client";
 import TablePagination from "@material-ui/core/TablePagination";
 import * as numbro from "numbro";
 import moment from "moment";